refactor: extract params handling logic to useParamsHandler hook and add loading state control
