+++ src/hooks/index.ts
@@ -1,1 +1,3 @@
 export * from './useLoading';
+export * from './useParamsHandler';
+
