+++ src/types/params.ts
@@ -5,6 +5,7 @@ import { UiStates } from '../enums';
 export interface BaseParams {
   clientId: string;
   redirectUri: string;
+  waitingForParams: boolean;
   apiKey: string;
   utm: string;
   entryState?: UiStates;
